Update thread.h
Added flag "THREAD_ISTMAIN" to indicate a tgroup's main thread.
Remember struct tgroup_t had a member thread_t * tg_main; which holds a reference to the main thread in the group. Therefore, the newly added flag will supplement this reference to the main thread in the thread group(tgroup). This will facilitate testing for the main thread of a tgroup simply by testing if the flag (THREAD_ISTMAIN) is set in a particular thread being queried.